requirements.txt
Django==2.2.16
mixer==7.1.2
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
sorl-thumbnail==12.7.0
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.7.2
python-dotenv==0.19.2
djoser==2.1.0
gunicorn==20.0.4
pytz==2021.1
sqlparse==0.4.1